requirements.txt
click==7.1.2
dataclasses==0.8
fastapi==0.63.0
feedgen==0.9.0
h11==0.12.0
Jinja2==2.11.3
lxml==4.6.2
markdown2==2.4.0
MarkupSafe==1.1.1
peewee==3.14.1
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
rsa==4.7.1
six==1.15.0
starlette==0.13.6
Telethon==1.19.5
typing-extensions==3.7.4.3
uvicorn==0.13.3
